[Rust Frontend] Extract shared chat types into `vllm-chat-types`
