+++ rust/Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "src/bench",
     "src/chat",
+    "src/chat-types",
     "src/cmd",
     "src/engine-core-client",
     "src/llm",
@@ -135,6 +136,7 @@ uuid = { version = "1.22.0", features = ["v4"] }
 validator = { version = "0.20.0", features = ["derive"] }
 vllm-bench = { path = "src/bench" }
 vllm-chat = { path = "src/chat" }
+vllm-chat-types = { path = "src/chat-types" }
 vllm-engine-core-client = { path = "src/engine-core-client" }
 vllm-llm = { path = "src/llm" }
 vllm-managed-engine = { path = "src/managed-engine" }
