+++ rust/src/chat/Cargo.toml
@@ -7,7 +7,6 @@ license.workspace = true
 [dependencies]
 anyhow.workspace = true
 asynk-strim-attr.workspace = true
-easy-ext.workspace = true
 futures.workspace = true
 half.workspace = true
 indexmap.workspace = true
@@ -30,6 +29,7 @@ tokio.workspace = true
 tracing.workspace = true
 trait-set.workspace = true
 uuid.workspace = true
+vllm-chat-types.workspace = true
 vllm-engine-core-client.workspace = true
 vllm-llm.workspace = true
 vllm-parser.workspace = true
