+++ rust/src/chat/src/renderer/deepseek_v4/encoding.rs
@@ -519,7 +519,7 @@ fn write_chat_content(out: &mut String, content: &ChatContent) -> Result<()> {
         ChatContent::Text(text) => out.push_str(text),
         ChatContent::Parts(parts) => {
             for part in parts {
-                out.push_str(part.as_text()?);
+                out.push_str(part.as_text().map_err(Error::UnsupportedMultimodalContent)?);
             }
         }
     }
