+++ rust/src/parser/Cargo.toml
@@ -9,10 +9,10 @@ test-util = []
 
 [dependencies]
 easy-ext.workspace = true
-serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 thiserror-ext.workspace = true
+vllm-chat-types.workspace = true
 vllm-tokenizer.workspace = true
 winnow.workspace = true
 xgrammar-structural-tag.workspace = true
